Creating user interface for admin and staff members

+++ accounts/views.py
@@ -3,7 +3,7 @@ from concurrent.futures._base import LOGGER
 
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.forms import UserCreationForm
-from django.contrib.auth.mixins import LoginRequiredMixin
+from django.contrib.auth.mixins import LoginRequiredMixin, PermissionRequiredMixin
 from django.contrib.auth.models import User, AnonymousUser
 from django.contrib.auth.views import LoginView, PasswordChangeView
 from django.core.exceptions import PermissionDenied
@@ -12,15 +12,13 @@ from django.db.transaction import atomic
 from django.forms import DateField, CharField, Textarea, NumberInput, ModelForm, IntegerField, Form
 from django.shortcuts import render
 from django.urls import reverse_lazy
-from django.views.generic import CreateView, DetailView, FormView, UpdateView, DeleteView
+from django.views.generic import CreateView, DetailView, FormView, UpdateView, DeleteView, ListView
 
 from accounts.forms import UserProductModelForm, OrderModelForm
 from accounts.models import Profile, Order, UserProduct, OrderProduct
 from store.models import Product
 from store.views import product
 
-
-# Create your views here.
 
 class SubmittableLoginView(LoginView):
     template_name = 'registration/login.html'
@@ -122,12 +120,28 @@ def remove_item_from_cart(request, pk):
 # @login_required
 def your_cart(request):
     if request.user.is_authenticated:
-        user_products = UserProduct.objects.filter(user=request.user)
+        user=request.user
     else:
-        user_products = UserProduct.objects.filter(user=request.session.session_key)
+        user=request.session.session_key
+    user_products_orig = UserProduct.objects.filter(user=user)
+    exceeds_number = ''
     total = 0
-    for user_product in user_products:
+    for user_product in user_products_orig:
+        if user_product.quantity > user_product.product.stock:
+            max_quantity = user_product.product.stock
+            user_product.quantity = max_quantity
+            user_product.save()
+            exceeds_number = user_product.product.title
         total += user_product.product.price * user_product.quantity
+        if user_product.quantity == 0:
+            user_product.delete()
+    user_products = UserProduct.objects.filter(user=user)
+    if exceeds_number != '':
+        notice_a = 'The demanded quantity for the last added product ('
+        notice_b = ') exceeded the stock and was changed accordingly!'
+        context = {'user_products': user_products, 'total': total, 'exceeds_number': exceeds_number,
+                   'notice_a': notice_a, 'notice_b': notice_b}
+        return render(request, 'cart.html', context)
     context = {'user_products': user_products, 'total': total}
     return render(request, 'cart.html', context)
 
@@ -230,14 +244,22 @@ def order_confirmation(request, pk):
 
 @ login_required
 def order_decline(request, pk):
+    user = request.user
     order = Order.objects.get(pk=pk)
+    if order.profile.user.id != user.id:
+        return your_orders(request)
+    if OrderProduct.objects.filter(order=order).exists():
+        return order_summary(request, pk)
     order.delete()
     return your_cart(request)
 
 
 @login_required
 def order_summary(request, pk):
+    user = request.user
     order = Order.objects.get(pk=pk)
+    if order.profile.user.id != user.id:
+        return your_orders(request)
     order_products = OrderProduct.objects.filter(order=order)
     payment_deadline = datetime.fromtimestamp(order.date_of_creation.timestamp() + 30*24*60*60)
     context = {'order': order, 'order_products': order_products, 'payment_deadline': payment_deadline}
@@ -249,3 +271,10 @@ def your_orders(request):
     orders = Order.objects.filter(profile__user=request.user)
     context = {'orders': orders}
     return render(request, 'orders.html', context)
+
+
+class OrdersListView(PermissionRequiredMixin, ListView):
+    template_name = "orders_overview.html"
+    model = Order
+    context_object_name = 'all_orders'
+    permission_required = 'accounts.view_order'
